+++ jestSetup.js
@@ -1,5 +1,21 @@
-global.environment = 'profserv02'
-// global.xAuthToken = ''
-global.channel = 'shared-english'
-global.path = 'pages/test-page'
-global.folderPath = '/content/documents/shared-english/pages'
+global.environment = process.env.ENVIRONMENT || ''
+global.xAuthToken = process.env.X_AUTH_TOKEN || ''
+
+// Site Management API Tests
+global.channelId = process.env.CHANNEL_ID || ''
+// global.componentGroup = process.env.COMPONENT_GROUP
+// global.component = process.env.COMPONENT
+// global.route = process.env.ROUTE
+// global.layout = process.env.LAYOUT
+// global.menu = process.env.MENU
+
+// Folder Management API Tests
+// global.folderPath = process.env.FOLDER_PATH
+
+// Content Management API Tests
+// global.documentPath = process.env.DOCUMENT_PATH
+// global.pagePath = process.env.PAGE_PATH
+// global.resourceBundlePath = process.env.RESOURCE_BUNDLE_PATH
+
+// Content Type Management API Tests
+// global.contentType = process.env.CONTENT_TYPE