+++ src/layouts/SidebarLayout/Header/index.js
@@ -14,7 +14,7 @@ import {
 import MenuTwoToneIcon from '@mui/icons-material/MenuTwoTone';
 import { SidebarContext } from 'src/contexts/SidebarContext';
 import CloseTwoToneIcon from '@mui/icons-material/CloseTwoTone';
-import HeaderMenu from './Menu';
+import Environments from './Environments'
 
 const HeaderWrapper = styled(Box)(
   ({ theme }) => `
@@ -65,7 +65,7 @@ function Header() {
         alignItems="center"
         spacing={2}
       >
-        {/* <HeaderMenu /> */}
+        {/* <Environments /> */}
       </Stack>
       <Box display="flex" alignItems="center">
         <Box