+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
+All notable changes to this project will be documented in below
 
-All notable changes to this project will be documented in this file
+### 0.1.1-alpha (April 18, 2023)
+- Fixed bug with links in Projects list
+- Added ability to modify channel details
+- Added loading icon to modal buttons when processing requests
+- Added TODO annotations for future development tasks
 
-## March 31, 2023
+### 0.1.0-alpha (March 31, 2023)
 - Update page title to reflect the name of the page
 - Added GETTING-STARTED.md into the application at /getting-started
 - Added CONTRIBUTING.md into the application at /contributing
 
-## March 30, 2023
+### March 30, 2023
 - Add release notes from CHANGELOG.md into the application at /release-notes